tests: modify to fit expected ggspec output
I've verified by running `make check` that the following changes are
acceptable to automake:

  - Print plan (number of tests) at end of output.

  - Don't print test number when printing each test.

  - Print diagnostics when a test fails.

My plan is to use my `ggspec` script, which is a test harness/runner, as
the single script to call from automake. Upon which it will traverse the
`spec` subdirectory, run all tests therein as usual, and output the test
results.

I'll implement the TAP output soon; after which I'll figure out how to
get automake to run ggspec.