tests: exit from each test file with failure status
The 'suite-failed' function returns the number of failed tests in the
suite. If no tests failed, exit with status code 0. Otherwise, with
status code equal to the number of tests that failed.

This tells Automake whether a test file failed or not; but it looks like
Guile 2's extra compilation messages is interfering with Automake's
parsing of ggspec's TAP-style output. That's my working theory, anyway.